src/requirements.txt
evdev==1.6.1
getmac==0.9.5
paho-mqtt==2.1.0
pexpect==4.8.0
ptyprocess==0.7.0
pycairo==1.23.0
PyGObject==3.42.2
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
scapy==2.5.0
six==1.16.0
